Reranking/requirements.txt
torch==2.1.2
transformers==4.34.0
icecream==2.1.3
tiktoken==0.5.2
mteb==1.1.1
